+++ backend/database.py
@@ -9,19 +9,16 @@ def construct_db_url(
     password: str = "1234",
     host: str = "localhost",
     port: str = "5432",
-    database: str = "db",
+    database: str = "db_dicom_market",
     dialect="postgresql",
     driver="psycopg2",
 ) -> str:
-
-    #return "sqlite:///./sql_app.db"
 
     return f"{dialect}://{user}:{password}@{host}:{port}/{database}"
 
 
 def get_connection():
 
-    # SQLALCHEMY_DATABASE_URL = "sqlite:///./sql_app.db"
     SQLALCHEMY_DATABASE_URL = construct_db_url()
     try:
         # The Engine is the starting point for any SQLAlchemy application. It’s “home base” for the actual database and its DBAPI
@@ -30,7 +27,7 @@ def get_connection():
         return engine
     except Exception as ex:
         print(
-            f"Connection could not be establiushed due to {ex} in url {SQLALCHEMY_DATABASE_URL}"
+            f"Connection could not be established due to {ex} in url {SQLALCHEMY_DATABASE_URL}"
         )
 
 
@@ -40,12 +37,3 @@ engine = get_connection()
 # each instance of this class will be a Session
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
-
-Base = declarative_base()
-
-try:
-    Base.metadata.create_all(bind=engine)  # creates the tables from db_models
-except Exception as ex:
-    print("An exception occured: ", ex)
-
-db = SessionLocal()