+++ backend/models.py
@@ -17,7 +17,7 @@ class User(Base):
     id = Column(Integer, primary_key=True)
     email = Column(String, unique=True, index=True)
     hashed_password = Column(String)
-    is_active = Column(Boolean, default=True)
+    is_active = Column(Boolean, default=True)  # tbd what is_actve is used for
 
     items = relationship("Item", back_populates="owner")
 
@@ -31,31 +31,3 @@ class Item(Base):
     owner_id = Column(Integer, ForeignKey("users.id"))
 
     owner = relationship("User", back_populates="items")
-"""
-
-class User(Base): #why have this inheritance
-    __tablename__ = "users"
-    id = Column(Integer, primary_key=True)
-    email = Column(String, unique=True, index=True)
-    pwd = Column(String)
-
-    items = relationship("Item", back_populates="owner")
-
-
-    def __repr__(self):
-        return f"User( id={self.id},email={self.email}, pwd={self.pwd}, items = {self.items})"
-
-
-class Item(Base):
-    __tablename__ = "items"
-    id = Column(Integer, primary_key=True)
-    title = Column(String, index=True)
-    description = Column(String, index=True)
-    owner_id = Column(Integer, ForeignKey("users.id"))
-
-    owner = relationship("User", back_populates="items")    
-
-    def __repr__(self):
-        return f"Item( id={self.id},title={self.title}, description={self.description}, owner = {self.owner_id})"
-
-"""