+++ frontend/src/Components/Header.js
@@ -1,9 +1,10 @@
-import React from "react";
+import React, { useState } from "react";
 import Navigation from "./Navigation";
 import { Link } from "react-router-dom";
 import { headerClass } from "../commonStyling";
 
-function Header() {
+function Header(props) {
+
 
     return (
         <header className={headerClass}>
@@ -13,8 +14,11 @@ function Header() {
             <span>The best DICOM in the market</span>
         
         
-        
+        <div className="flex flex-row w-60 bg-blue-100 justify-between">
+            <div className="pr-5">{props.user}</div>
         <Navigation />
+
+        </div>
         </header>
     )