+++ CHANGELOG.md
@@ -13,3 +13,7 @@
 26 Sep 2024
 - Backend: Connection with postgreSQL and create user
 - Frontend: Factoring out styling and new Signup page component/route
+
+28 Sep 2024
+- Backend: Finished apis to manage users (CRUD) & better error handling (HTTP exceptions)
+- Frontent: Connection with Backend api to register new user & UserProfilePage placeholder site